pyproject.toml
[project]
name = "eval-steering"
version = "0.1.0"
requires-python = ">=3.10"
dependencies = [
    "torch",
    "transformers",
    "accelerate",
    "openai",
    "pyyaml",
    "python-dotenv",
    "tqdm",
]
